fix: restructure (separate pages and components)
